zelta_ai/requirements.txt
# --- Core API & Server ---
fastapi==0.110.1
uvicorn[standard]==0.29.0

# --- Live Data & HTTP ---
aiohttp==3.9.3
httpx==0.27.0
beautifulsoup4==4.12.3
websockets==12.0

# --- Quantitative Math & Statistics ---
numpy==1.26.4
scipy==1.13.0

# --- Natural Language Processing (NLP) ---
transformers==4.39.3
torch==2.2.2

# --- AI & Cloud Integrations (ZELTA Co-Pilot) ---
google-generativeai==0.4.1
google-cloud-aiplatform==1.46.0

# --- Testing (Development Only) ---
pytest==8.1.1
pytest-asyncio==0.23.6